Conflicts:
	Makefile.am
	bin/data/Ruleset/Xcom1Ruleset.rul
	src/Battlescape/ActionMenuItem.cpp
	src/Battlescape/ActionMenuItem.h
	src/Battlescape/AliensCrashState.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/InfoboxOKState.cpp
	src/Battlescape/InfoboxState.cpp
	src/Battlescape/Inventory.cpp
	src/Battlescape/PrimeGrenadeState.cpp
	src/CMakeLists.txt
	src/Engine/State.cpp
	src/Geoscape/Globe.cpp
	src/Geoscape/Globe.h
	src/Interface/Frame.cpp
	src/Interface/Text.cpp
	src/Interface/Text.h
	src/OpenXcom.2010.vcxproj
	src/Resource/ResourcePack.cpp
	src/Resource/ResourcePack.h
	src/Resource/XcomResourcePack.cpp
	src/Ruleset/Polygon.cpp
	src/Ruleset/Polygon.h
	src/Ruleset/Polyline.cpp
	src/Ruleset/Polyline.h
	src/Ruleset/Ruleset.cpp
	src/Ruleset/Ruleset.h
	src/Savegame/SavedGame.cpp
	src/Savegame/SavedGame.h
	src/pch.h

+++ src/Basescape/PurchaseState.cpp
@@ -867,7 +867,7 @@ void PurchaseState::increaseByValue(int change)
 	}
 	else
 	{
-		int maxByMoney = (_game->getSavedGame()->getFunds() - _total) / getPrice();
+		int maxByMoney = (static_cast<int>(_game->getSavedGame()->getFunds()) - _total) / getPrice(); // note: (int)cast renders int64_t useless.
 		change = std::min(maxByMoney, change);
 
 		if (_sel <= 2) // Personnel count
